🔨 refactor: organize code in a clean and maintainable way

+++ src/utils/index.ts
@@ -2,18 +2,20 @@ import { ensureDir, exists } from 'https://deno.land/std@0.224.0/fs/mod.ts'
 import { Expense } from '../models/expense.ts'
 
 const dataFolder = 'data/'
-// create an array to store the expenses
-const expenses: Expense[] = []
 
 export async function loadExpencesFromCSV() {
-  // Ensure data folder exists and create the file if it doesn't
   await setupDataFolder()
-  const fileExists = await exists(`${dataFolder}/expenses.csv`)
-  if (!fileExists) {
-    await Deno.writeTextFile(`${dataFolder}/expenses.csv`, '')
+  try {
+    const fileExists = await exists(`${dataFolder}/expenses.csv`, { isReadable: true })
+    if (!fileExists) await Deno.writeTextFile(`${dataFolder}/expenses.csv`, '')
+  } catch (error) {
+    if (!(error instanceof Deno.errors.NotFound)) {
+      throw error
+    }
   }
   const csvContent: string = await Deno.readTextFile(`${dataFolder}/expenses.csv`)
   const lines: string[] = csvContent.split('\n')
+  const expenses: Expense[] = [] // Move expenses array inside the function
   lines.forEach((line: string) => {
     if (line.trim() === '') return // Skip empty lines
     const [id, description, amount, date] = line.split(',')
@@ -31,29 +33,24 @@ export async function loadExpencesFromCSV() {
   })
   return expenses
 }
-
 async function setupDataFolder() {
-  await ensureDir(dataFolder)
+  try {
+    await ensureDir(dataFolder)
+  } catch (error) {
+    if (!(error instanceof Deno.errors.AlreadyExists)) {
+      throw error
+    }
+  }
 
-  if (!(await exists(`${dataFolder}/expenses.csv`))) {
+  if (!(await exists(`${dataFolder}/expenses.csv`, { isReadable: true }))) {
     await Deno.writeTextFile(`${dataFolder}/expenses.csv`, '')
   }
 }
 
-export async function saveExpencesToCSV() {
+export async function saveExpencesToCSV(expenses: Expense[]) {
   await setupDataFolder()
   const csvContent = expenses
     .map((expense) => `${expense.id},${expense.description},${expense.amount},${expense.date}`)
     .join('\n')
   await Deno.writeTextFile(`${dataFolder}/expenses.csv`, csvContent)
-}
-
-export async function listExpenses() {
-  const expenses = await loadExpencesFromCSV()
-  console.log('Expenses:')
-  expenses.forEach((expense) => {
-    console.log(
-      `ID: ${expense.id}, Description: ${expense.description}, Amount: ${expense.amount}, Date: ${expense.date}`
-    )
-  })
 }
